+++ app/presenters/stage_presenter.rb
@@ -1,0 +1,9 @@
+class StagePresenter < SimpleDelegator
+  def name
+    stage_name
+  end
+
+  def actions
+    action_states.map { |action| ActionPresenter.new(action) }
+  end
+end