+++ app/services/create_infrastructure_variable.rb
@@ -16,7 +16,7 @@ class CreateInfrastructureVariable
   private
 
   def aws_ssm_client
-    ClientForCoreAwsAccount.new.call
+    SSMClientForCoreAwsAccount.new.call
   end
 
   def name_with_path